Add notification actions/tests
