+++ client/control/tests/actions/test_controlActions.js
@@ -5,6 +5,10 @@ import fetchMock from 'fetch-mock';
 import makeApiRequest from 'control/api';
 import * as actionTypes from 'control/actions/ControlActions';
 
+import {
+  SHOW_NOTIFICATION,
+} from 'control/actions/NotificationActions';
+
 import { fixtureRecipes, initialState } from 'control/tests/fixtures';
 
 const middlewares = [thunk];
@@ -51,7 +55,7 @@ describe('controlApp Actions', () => {
 
     it('creates a SHOW_NOTIFICATION action if provided', () => {
       const expectedAction = {
-        type: actionTypes.SHOW_NOTIFICATION,
+        type: SHOW_NOTIFICATION,
         notification: {
           messageType: 'error',
           message: 'Error fetching recipes.',
@@ -63,22 +67,6 @@ describe('controlApp Actions', () => {
       return store.dispatch(makeApiRequest('fetchAllRecipes')).catch(() => {
         expect(store.getActions()).toContain(expectedAction);
       });
-    });
-  });
-
-  describe('showNotification', () => {
-    it('automatically dismisses notifications after 10 seconds', async () => {
-      jasmine.clock().install();
-
-      const notification = { messageType: 'success', message: 'message' };
-      await store.dispatch(actionTypes.showNotification(notification));
-
-      const dismissAction = actionTypes.dismissNotification(notification.id);
-      expect(store.getActions()).not.toContain(dismissAction);
-      jasmine.clock().tick(10001);
-      expect(store.getActions()).toContain(dismissAction);
-
-      jasmine.clock().uninstall();
     });
   });
 });
